addressed comments
