+++ src/eva/vision/data/wsi/patching/coordinates.py
@@ -4,9 +4,12 @@ import dataclasses
 import functools
 from typing import List, Tuple
 
+import cv2
+import numpy as np
+
 from eva.vision.data.wsi import backends
+from eva.vision.data.wsi.backends.base import Wsi
 from eva.vision.data.wsi.patching import samplers
-from eva.vision.utils.io import get_mask
 
 LRU_CACHE_SIZE = 32
 
@@ -55,17 +58,15 @@ class PatchCoordinates:
         mpp_ratio = target_mpp / level_mpp
         scaled_width, scaled_height = int(mpp_ratio * width), int(mpp_ratio * height)
 
-        sample_args = {}
-        sample_args["width"] = scaled_width
-        sample_args["height"] = scaled_height
-        sample_args["layer_shape"] = wsi.level_dimensions[level_idx]
-
+        sample_args = {
+            "width": scaled_width,
+            "height": scaled_height,
+            "layer_shape": wsi.level_dimensions[level_idx],
+        }
         if isinstance(sampler, samplers.ForegroundSampler):
-            sample_args["mask"] = get_mask(wsi, wsi_path, level_idx)
+            sample_args["mask"] = get_mask(wsi, level_idx)
 
-        x_y = []
-        for x, y in sampler.sample(**sample_args):
-            x_y.append((x, y))
+        x_y = [(x, y) for x, y in sampler.sample(**sample_args)]
 
         return cls(x_y, scaled_width, scaled_height, level_idx)
 
@@ -88,3 +89,38 @@ def get_cached_coords(
         backend=backend,
         sampler=sampler,
     )
+
+
+def get_mask(
+    wsi: Wsi,
+    level_idx: int,
+    kernel_size: tuple[int, int] = (7, 7),
+    gray_threshold: int = 220,
+    fill_holes: bool = False,
+) -> tuple[np.ndarray, float]:
+    """Extracts a binary mask from an image.
+
+    Args:
+        wsi: The WSI object.
+        level_idx: The level index to extract the mask from.
+        kernel_size: The size of the kernel for morphological operations.
+        gray_threshold: The threshold for the gray scale image.
+        fill_holes: Whether to fill holes in the mask.
+    """
+    image = np.array(
+        wsi.read_region([0, 0], len(wsi.level_dimensions) - 1, wsi.level_dimensions[-1])
+    )
+
+    kernel = cv2.getStructuringElement(cv2.MORPH_ELLIPSE, kernel_size)
+    gray = cv2.cvtColor(image, cv2.COLOR_RGB2GRAY)
+    mask = np.where(gray < gray_threshold, 1, 0).astype(np.uint8)
+
+    if fill_holes:
+        mask = cv2.dilate(mask, kernel, iterations=1)
+        contour, _ = cv2.findContours(mask, cv2.RETR_CCOMP, cv2.CHAIN_APPROX_SIMPLE)
+        for cnt in contour:
+            cv2.drawContours(mask, [cnt], 0, 1, -1)
+
+    mask_scale_factor = wsi.level_dimensions[-1][0] / wsi.level_dimensions[level_idx][0]
+
+    return mask, mask_scale_factor
