+++ src/eva/vision/utils/io/__init__.py
@@ -1,12 +1,11 @@
 """Vision I/O utilities."""
 
-from eva.vision.utils.io.image import get_mask, read_image
+from eva.vision.utils.io.image import read_image
 from eva.vision.utils.io.nifti import fetch_total_nifti_slices, read_nifti_slice
 from eva.vision.utils.io.text import read_csv
 
 __all__ = [
     "read_image",
-    "get_mask",
     "fetch_total_nifti_slices",
     "read_nifti_slice",
     "read_csv",
